Organize test suites into unit and integration
This is the goals of project organization composing different
combinations to run any number of tests together.
