+++ tests/unit/Support/Calendar/Periodicity/QuarterlyTest.php
@@ -19,32 +19,37 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace Tests\Support\Calendar\Periodicity;
+namespace Tests\unit\Support\Calendar\Periodicity;
 
 use Carbon\Carbon;
 use FireflyIII\Support\Calendar\Periodicity;
 use FireflyIII\Support\Calendar\Periodicity\Interval;
 
-class HalfYearlyTest extends IntervalTestCase
+/**
+ * @group unit-test
+ * @group support
+ * @group calendar
+ * @group periodicity
+ */
+class QuarterlyTest extends IntervalTestCase
 {
     public static function factory(): Interval
     {
-        return new Periodicity\HalfYearly();
+        return new Periodicity\Quarterly();
     }
 
     public static function provideIntervals(): array
     {
         return [
-            new IntervalProvider(Carbon::now(), Carbon::now()->addMonthsNoOverflow(6)),
-            new IntervalProvider(Carbon::parse('2019-01-29'), Carbon::parse('2019-07-29')),
-            new IntervalProvider(Carbon::parse('2019-01-30'), Carbon::parse('2019-07-30')),
-            new IntervalProvider(Carbon::parse('2019-01-31'), Carbon::parse('2019-07-31')),
-            new IntervalProvider(Carbon::parse('2018-11-01'), Carbon::parse('2019-05-01')),
-            new IntervalProvider(Carbon::parse('2019-08-29'), Carbon::parse('2020-02-29')),
-            new IntervalProvider(Carbon::parse('2019-08-30'), Carbon::parse('2020-02-29')),
-            new IntervalProvider(Carbon::parse('2019-08-31'), Carbon::parse('2020-02-29')),
-            new IntervalProvider(Carbon::parse('2020-08-29'), Carbon::parse('2021-02-28')),
-            new IntervalProvider(Carbon::parse('2020-08-30'), Carbon::parse('2021-02-28')),
+            new IntervalProvider(Carbon::now(), Carbon::now()->addMonths(3)),
+            new IntervalProvider(Carbon::parse('2019-01-29'), Carbon::parse('2019-04-29')),
+            new IntervalProvider(Carbon::parse('2019-01-30'), Carbon::parse('2019-04-30')),
+            new IntervalProvider(Carbon::parse('2019-01-31'), Carbon::parse('2019-04-30')),
+            new IntervalProvider(Carbon::parse('2018-11-01'), Carbon::parse('2019-02-01')),
+            new IntervalProvider(Carbon::parse('2019-11-29'), Carbon::parse('2020-02-29')),
+            new IntervalProvider(Carbon::parse('2019-11-30'), Carbon::parse('2020-02-29')),
+            new IntervalProvider(Carbon::parse('2020-11-29'), Carbon::parse('2021-02-28')),
+            new IntervalProvider(Carbon::parse('2020-11-30'), Carbon::parse('2021-02-28')),
         ];
     }
 }
