+++ tests/unit/Support/Calendar/Periodicity/YearlyTest.php
@@ -19,24 +19,31 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace Tests\Support\Calendar\Periodicity;
+namespace Tests\unit\Support\Calendar\Periodicity;
 
 use Carbon\Carbon;
 use FireflyIII\Support\Calendar\Periodicity;
 use FireflyIII\Support\Calendar\Periodicity\Interval;
 
-class WeeklyTest extends IntervalTestCase
+/**
+ * @group unit-test
+ * @group support
+ * @group calendar
+ * @group periodicity
+ */
+class YearlyTest extends IntervalTestCase
 {
     public static function factory(): Interval
     {
-        return new Periodicity\Weekly();
+        return new Periodicity\Yearly();
     }
 
     public static function provideIntervals(): array
     {
         return [
-            new IntervalProvider(Carbon::now(), Carbon::now()->addWeek()),
-            new IntervalProvider(Carbon::parse('2023-01-31'), Carbon::parse('2023-02-07')),
+            new IntervalProvider(Carbon::now(), Carbon::now()->addYears(1)),
+            new IntervalProvider(Carbon::parse('2019-01-29'), Carbon::parse('2020-01-29')),
+            new IntervalProvider(Carbon::parse('2020-02-29'), Carbon::parse('2021-02-28')),
         ];
     }
 }
